Reorganize some of the integration tests.
